Minor QoL improvements to TextDisplay, quick 'ready' event fix and README update
'ready' event used to be hardcoded to emit 1 millisecond after we
initiate connection. This, in every single aspects, was not a good idea.
But it worked.

After actually bothering to look at the node `serialport` library, I
added a passthrough 'read' event without the need for hardcoding.

You're welcome, past me.